+++ tests/createCustomizePoolWConfig.test.ts
@@ -172,8 +172,7 @@ describe("Initialize customizable pool with dynamic config", () => {
               startingFeeBps: 5000,
               endingFeeBps: 100,
               numberOfPeriod: 180,
-              startingMarketCap: 20_000,
-              endingMarketCap: 20_000_000,
+              priceMultiple: 1000, // e.g. $20k → $20M market cap
               schedulerExpirationDuration: 2592000, // 30 days
             },
           },
@@ -247,8 +246,7 @@ describe("Initialize customizable pool with dynamic config", () => {
               startingFeeBps: 5000,
               endingFeeBps: 100,
               numberOfPeriod: 180,
-              startingMarketCap: 20_000,
-              endingMarketCap: 20_000_000,
+              priceMultiple: 1000, // e.g. $20k → $20M market cap
               schedulerExpirationDuration: 2592000, // 30 days
             },
           },